+++ R/Functions.R
@@ -261,7 +261,7 @@ pairwiseMutions <- function(germ_imgt,
     }
     ##### count informative positions
     if (norm_fact) {
-        informative_pos <- sapply(1:n, function(x){ sum(stri_count(seq_imgt[x], fixed = c("A","C","G","T"))) })   
+        informative_pos <- sapply(1:n, function(x){ sum(stringi::stri_count(seq_imgt[x], fixed = c("A","C","G","T"))) })   
     } else {
         informative_pos <- rep(1, n)
     }
@@ -372,7 +372,7 @@ calculateInterVsIntra <- function(db,
                       .errorhandling='stop') %dopar% {
                           
                           # *********************************************************************************
-                          clones <- stri_split_fixed(vjl_gps$clone_id[k], ",")[[1]]
+                          clones <- stringi::stri_split_fixed(vjl_gps$clone_id[k], ",")[[1]]
                           l <- vjl_gps$junction_length[k]
                           n_clones <- length(clones)
                           in_clones <- db[[clone]] %in% clones
@@ -425,7 +425,7 @@ calculateInterVsIntra <- function(db,
                          row.names=NULL)
     db_dff$label <- "intra"
     db_dff$label[grepl("inter", db_dff$keyName)] <- "inter"
-    clones_xy <- data.frame(matrix(unlist(stri_split_fixed(db_dff$keyName, "_", n=3)),
+    clones_xy <- data.frame(matrix(unlist(stringi::stri_split_fixed(db_dff$keyName, "_", n=3)),
                                    nrow=nrow(db_dff),
                                    byrow=T),
                             stringsAsFactors=FALSE)
@@ -482,6 +482,15 @@ prepare_db <- function(db,
                        first = FALSE, cdr3 = FALSE, fields = NULL,
                        cell_id = NULL, locus = NULL, only_heavy = TRUE,
                        mod3 = FALSE, max_n = 0) {
+    #TODO: remove only_heavy parameter when it becomes deprecated.
+    if(!only_heavy){
+      warning("The only_heavy = FALSE parameter is deprecated. Will run as if only_heavy = TRUE")
+      only_heavy <- TRUE
+    }
+  
+    #TODO: double check that the 'junction_l' check are still needed when only_heavy
+    # is fully removed 
+
     # add junction length column
     db$junction_l <- stringi::stri_length(db[[junction]])
     junction_l <- "junction_l"
@@ -489,9 +498,13 @@ prepare_db <- function(db,
     ### check for mod3
     # filter mod 3 junction lengths
     if (mod3) {
-        n_rmv_mod3 <- sum(db[[junction_l]]%%3 != 0)
+        n_before <- nrow(db)
         db <- db %>% 
             dplyr::filter(!!rlang::sym(junction_l)%%3 == 0)
+        n_after <- nrow(db)
+        if ( n_before > n_after) {
+            warning(paste("Removed", n_before - n_after, "sequences with junction length not divisible by 3."))
+        }
     } else {
         n_rmv_mod3 <- 0
     }
@@ -502,6 +515,9 @@ prepare_db <- function(db,
         n_rmv_cdr3 <- sum(db[[junction_l]] <= 6)
         db <- db %>% 
             dplyr::filter(!!rlang::sym(junction_l) > 6)
+        if ( n_rmv_cdr3 > 0) {
+            warning(paste("Removed", n_rmv_cdr3, "sequences with junction length <=6."))
+        }            
         # add cdr3 column
         db$cdr3_col <- substr(db[[junction]], 4, db[[junction_l]]-3)
         cdr3_col <- "cdr3_col"
@@ -512,47 +528,56 @@ prepare_db <- function(db,
     
     ### check for degenerate characters (non-ATCG's)
     # Count the number of non-ATCG's in junction
+    # TODO: remove this from here as groupGenes already checks for ambiguous positions
     if (!is.null(max_n)) {
-        n_rmv_N <- sum(stri_count(db[[junction]], regex = "[^ATCG]") > max_n)
+        n_rmv_N <- sum(stringi::stri_count(db[[junction]], regex = "[^ATCG]") > max_n)
+        n_before <- nrow(db)
         db <- db %>% 
-            dplyr::filter(stri_count(!!rlang::sym(junction), regex = "[^ATCG]") <= max_n)
+            dplyr::filter(stringi::stri_count(!!rlang::sym(junction), regex = "[^ATCG]") <= max_n)
+        n_after <- nrow(db)
+        if ( n_before > n_after) {
+            warning(paste("Removed", n_before - n_after, "sequences with non ATCG charachters."))
+        }
     } else {
         n_rmv_N <- 0
     }
     
-    ### Parse V and J columns to get gene
+    ### Parse V and J columns to get gene groups (vj_group)
+    ### Within "fields" group, group sequences by V and J calls.
     if (!is.null(fields)) {
         . <- NULL
         db <- db %>%
             dplyr::group_by(!!!rlang::syms(fields)) %>%
-            do(groupGenes(., 
+            do(alakazam::groupGenes(., 
                           v_call = v_call,
                           j_call = j_call,
-                          junc_len = NULL,
+                          junc_len = NULL, #TODO: junc_len could be not NULL, why not passing it?
                           cell_id = cell_id,
                           locus = locus,
-                          only_heavy = only_heavy,
+                          only_heavy = T, #TODO: we only allow True for now, when deprecated remove.
                           first = first))        
     } else {
-        db <- groupGenes(db,
+        db <- alakazam::groupGenes(db,
                          v_call = v_call,
                          j_call = j_call,
-                         junc_len = NULL,
+                         junc_len = NULL, #TODO: junc_len could be not NULL, why not passing it?
                          cell_id = cell_id,
                          locus = locus,
-                         only_heavy = only_heavy,
+                         only_heavy = T, #TODO: we only allow True for now, when deprecated remove.
                          first = first)        
     }
     
-    ### groups to use
+    
+    ### vj_group comes from groupGenes
     groupBy <- c("vj_group", junction_l, fields)
     
     ### assign group ids to db
+    ### TODO: why does it split here by junction and not within groupGenes
     db$vjl_group <- db %>%
         dplyr::group_by(!!!rlang::syms(groupBy)) %>%
         dplyr::group_indices()
     
-    ### retrun results
+    ### return results
     return_list <- list("db" = db, 
                         "n_rmv_mod3" = n_rmv_mod3, 
                         "n_rmv_cdr3" = n_rmv_cdr3,
@@ -603,7 +628,7 @@ pairwiseMutMatrix <- function(informative_pos, mutMtx, motifMtx) {
 #'
 #' @examples
 #' # Find clones
-#' results <- hierarchicalClones(ExampleDb, threshold=0.15)
+#' results <- hierarchicalClones(ExampleDb, threshold=0.15, summarize_clones=TRUE)
 #' 
 #' # Plot clonal summaries
 #' plot(results, binwidth=0.02)
@@ -612,6 +637,16 @@ pairwiseMutMatrix <- function(informative_pos, mutMtx, motifMtx) {
 plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL, 
                              binwidth=NULL, title=NULL, size=0.75, silent=FALSE, 
                              ...) {
+    
+    # Check if data is of ScoperClones class
+    if (!inherits(data, "ScoperClones")) {
+        warning("Input data is not of class 'ScoperClones'. ",
+                "Please ensure you are using the output from hierarchicalClones(), ",
+                "identicalClones(), or spectralClones() functions with ",
+                "summarize_clones=TRUE.",
+                call. = FALSE)
+        stop("Invalid input data class")
+    }
     
     eff_threshold <- data@eff_threshold
     xdf <- select(data@inter_intra, c("distance", "label"))
@@ -656,7 +691,7 @@ plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL,
                            binwidth = binwidth, color = "white", alpha = 0.85) +
             geom_density(data = data_intra, 
                          aes(x = !!rlang::sym("distance")),
-                         size = size, color = "grey30")
+                         linewidth = size, color = "grey30")
     }
     
     # Plot between clonal distances
@@ -669,7 +704,7 @@ plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL,
                            binwidth = binwidth, color = "white", alpha = 0.75) +
             geom_density(data = data_inter, 
                          aes(x = !!rlang::sym("distance")),
-                         size = size, color = "grey60")
+                         linewidth = size, color = "grey60")
     } else {
         warning("No inter clonal distance is detected. Each group of sequences with same V-gene, J-gene, and junction length may contain only one clone.")
     }
@@ -678,7 +713,7 @@ plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL,
     if (!is.na(eff_threshold)) {
         p <- p + 
             ggtitle(paste("Effective threshold=", eff_threshold)) +
-            geom_vline(xintercept=eff_threshold, color="grey30", linetype=2, size=size)
+            geom_vline(xintercept=eff_threshold, color="grey30", linetype=2, linewidth=size)
     } else {
         p <- p + 
             ggtitle(paste("Effective threshold not found"))
@@ -739,10 +774,12 @@ plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL,
 #' @param    locus              name of the column containing locus information. 
 #'                              Only applicable to single-cell data.
 #'                              Ignored if \code{cell_id=NULL}.
-#' @param    only_heavy         use only the IGH (BCR) or TRB/TRD (TCR) sequences 
+#' @param    only_heavy         This is deprecated. Only heavy chains will be used in clustering.
+#'                              Use only the IGH (BCR) or TRB/TRD (TCR) sequences 
 #'                              for grouping. Only applicable to single-cell data.
 #'                              Ignored if \code{cell_id=NULL}.
-#' @param    split_light        split clones by light chains. Ignored if \code{cell_id=NULL}.
+#' @param    split_light        This is deprecated. If you desire to split clones by light chains 
+#'                              use dowser::resolveLightChains.
 #' @param    first              specifies how to handle multiple V(D)J assignments for initial grouping. 
 #'                              If \code{TRUE} only the first call of the gene assignments is used. 
 #'                              If \code{FALSE} the union of ambiguous gene assignments is used to 
@@ -763,16 +800,16 @@ plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL,
 #'                              The input file directory is used if path is not specified.
 #'                              The default is \code{NULL} for no action.
 #' @param    summarize_clones   if \code{TRUE} performs a series of analysis to assess the clonal landscape
-#'                              and returns a \link{ScoperClones} object. If \code{FALSE} then
+#'                              and returns a \link{ScoperClones} object. If \code{FALSE} (default) then
 #'                              a modified input \code{db} is returned. When grouping by \code{fields}, 
-#'                              \code{summarize_clones} should be \code{FALSE}. 
-#'
+#'                              \code{summarize_clones} should be \code{FALSE}.
+#' 
 #' @return
-#' If \code{summarize_clones=TRUE} (default) a \link{ScoperClones} object is returned that includes the 
+#' If \code{summarize_clones=FALSE} (default) a modified \code{data.frame} is returned with clone identifiers in the 
+#' specified \code{clone} column.
+#' If \code{summarize_clones=TRUE} a \link{ScoperClones} object is returned that includes the 
 #' clonal assignment summary information and a modified input \code{db} in the \code{db} slot that 
 #' contains clonal identifiers in the specified \code{clone} column.
-#' If \code{summarize_clones=FALSE} modified \code{data.frame} is returned with clone identifiers in the 
-#' specified \code{clone} column.
 #' 
 #' @section Single-cell data:
 #' To invoke single-cell mode the \code{cell_id} argument must be specified and the \code{locus} 
@@ -800,7 +837,7 @@ plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL,
 #'
 #' @examples
 #' # Find clonal groups
-#' results <- identicalClones(ExampleDb)
+#' results <- identicalClones(ExampleDb, summarize_clones=TRUE)
 #' 
 #' # Retrieve modified input data with clonal clustering identifiers
 #' df <- as.data.frame(results)
@@ -811,18 +848,16 @@ plotCloneSummary <- function(data, xmin=NULL, xmax=NULL, breaks=NULL,
 #' @export
 identicalClones <- function(db, method=c("nt", "aa"), junction="junction", 
                             v_call="v_call", j_call="j_call", clone="clone_id", fields=NULL,
-                            cell_id=NULL, locus="locus", only_heavy=TRUE, split_light=TRUE,
+                            cell_id=NULL, locus="locus", only_heavy=TRUE, split_light=FALSE,
                             first=FALSE, cdr3=FALSE, mod3=FALSE, max_n=0, nproc=1,
-                            verbose=FALSE, log=NULL, 
-                            summarize_clones=TRUE) {
-    
+                            verbose=FALSE, log=NULL, summarize_clones=FALSE) {
+
     results <- defineClonesScoper(db = db,
                                   method = match.arg(method), model = "identical", 
                                   junction = junction, v_call = v_call, j_call = j_call, clone = clone, fields = fields,
                                   cell_id = cell_id, locus = locus, only_heavy = only_heavy, split_light = split_light,
                                   first = first, cdr3 = cdr3, mod3 = mod3, max_n = max_n, nproc = nproc,        
-                                  verbose = verbose, log = log, 
-                                  summarize_clones = summarize_clones)
+                                  verbose = verbose, log = log, summarize_clones = summarize_clones)
     
     ### return results
     if (summarize_clones) {
@@ -872,10 +907,12 @@ identicalClones <- function(db, method=c("nt", "aa"), junction="junction",
 #' @param    locus              name of the column containing locus information. 
 #'                              Only applicable to single-cell data.
 #'                              Ignored if \code{cell_id=NULL}.
-#' @param    only_heavy         use only the IGH (BCR) or TRB/TRD (TCR) sequences 
+#' @param    only_heavy         This is deprecated. Only heavy chains will be used in clustering.
+#'                              Use only the IGH (BCR) or TRB/TRD (TCR) sequences 
 #'                              for grouping. Only applicable to single-cell data.
 #'                              Ignored if \code{cell_id=NULL}.
-#' @param    split_light        split clones by light chains. Ignored if \code{cell_id=NULL}.
+#' @param    split_light        This is deprecated. If you desire to split clones by light chains 
+#'                              use dowser::resolveLightChains.
 #' @param    first              specifies how to handle multiple V(D)J assignments for initial grouping. 
 #'                              If \code{TRUE} only the first call of the gene assignments is used. 
 #'                              If \code{FALSE} the union of ambiguous gene assignments is used to 
@@ -898,16 +935,17 @@ identicalClones <- function(db, method=c("nt", "aa"), junction="junction",
 #'                              The input file directory is used if path is not specified.
 #'                              The default is \code{NULL} for no action.
 #' @param    summarize_clones   if \code{TRUE} performs a series of analysis to assess the clonal landscape
-#'                              and returns a \link{ScoperClones} object. If \code{FALSE} then
+#'                              and returns a \link{ScoperClones} object. If \code{FALSE} (default) then
 #'                              a modified input \code{db} is returned. When grouping by \code{fields}, 
-#'                              \code{summarize_clones} should be \code{FALSE}.
+#'                              \code{summarize_clones} should be \code{FALSE}. 
+#' @param   seq_id              The column containing sequence ids
 #'
 #' @return
-#' If \code{summarize_clones=TRUE} (default) a \link{ScoperClones} object is returned that includes the 
+#' If \code{summarize_clones=FALSE} (default) a modified \code{data.frame} is returned with clone identifiers in the 
+#' specified \code{clone} column.
+#' If \code{summarize_clones=TRUE} a \link{ScoperClones} object is returned that includes the 
 #' clonal assignment summary information and a modified input \code{db} in the \code{db} slot that 
 #' contains clonal identifiers in the specified \code{clone} column.
-#' If \code{summarize_clones=FALSE} modified \code{data.frame} is returned with clone identifiers in the 
-#' specified \code{clone} column.
 #' 
 #' @section Single-cell data:
 #' To invoke single-cell mode the \code{cell_id} argument must be specified and the \code{locus} 
@@ -936,7 +974,7 @@ identicalClones <- function(db, method=c("nt", "aa"), junction="junction",
 #'
 #' @examples
 #' # Find clonal groups
-#' results <- hierarchicalClones(ExampleDb, threshold=0.15)
+#' results <- hierarchicalClones(ExampleDb, threshold=0.15, summarize_clones=TRUE)
 #' 
 #' # Retrieve modified input data with clonal clustering identifiers
 #' df <- as.data.frame(results)
@@ -948,18 +986,16 @@ identicalClones <- function(db, method=c("nt", "aa"), junction="junction",
 hierarchicalClones <- function(db, threshold, method=c("nt", "aa"), linkage=c("single", "average", "complete"), 
                                normalize=c("len", "none"), junction="junction", 
                                v_call="v_call", j_call="j_call", clone="clone_id", fields=NULL,
-                               cell_id=NULL, locus="locus", only_heavy=TRUE, split_light=TRUE,
+                               cell_id=NULL, locus="locus", only_heavy=TRUE, split_light=FALSE,
                                first=FALSE, cdr3=FALSE, mod3=FALSE, max_n=0, nproc=1,
-                               verbose=FALSE, log=NULL,
-                               summarize_clones=TRUE) {
+                               verbose=FALSE, log=NULL, summarize_clones=FALSE, seq_id = "sequence_id") {
     
     results <- defineClonesScoper(db = db, threshold = threshold, model = "hierarchical", 
                                   method = match.arg(method), linkage = match.arg(linkage), normalize = match.arg(normalize),
                                   junction = junction, v_call = v_call, j_call = j_call, clone = clone, fields = fields,
                                   cell_id = cell_id, locus = locus, only_heavy = only_heavy, split_light = split_light,
                                   first = first, cdr3 = cdr3, mod3 = mod3, max_n = max_n, nproc = nproc,   
-                                  verbose = verbose, log = log, 
-                                  summarize_clones = summarize_clones)
+                                  verbose = verbose, log = log, summarize_clones = summarize_clones)
     
     # return results
     if (summarize_clones) {
@@ -1005,10 +1041,12 @@ hierarchicalClones <- function(db, threshold, method=c("nt", "aa"), linkage=c("s
 #' @param    locus              name of the column containing locus information. 
 #'                              Only applicable to single-cell data.
 #'                              Ignored if \code{cell_id=NULL}.
-#' @param    only_heavy         use only the IGH (BCR) or TRB/TRD (TCR) sequences 
+#' @param    only_heavy         This is deprecated. Only heavy chains will be used in clustering.
+#'                              Use only the IGH (BCR) or TRB/TRD (TCR) sequences 
 #'                              for grouping. Only applicable to single-cell data.
 #'                              Ignored if \code{cell_id=NULL}.
-#' @param    split_light        split clones by light chains. Ignored if \code{cell_id=NULL}.
+#' @param    split_light        This is deprecated. If you desire to split clones by light chains 
+#'                              use dowser::resolveLightChains.
 #' @param    targeting_model    \link[shazam]{TargetingModel} object. Only applicable if 
 #'                              \code{method="vj"}. See Details for description.
 #' @param    len_limit          \link[shazam]{IMGT_V} object defining the regions and boundaries of the Ig 
@@ -1039,16 +1077,15 @@ hierarchicalClones <- function(db, threshold, method=c("nt", "aa"), linkage=c("s
 #'                              The input file directory is used if path is not specified.
 #'                              The default is \code{NULL} for no action.
 #' @param    summarize_clones   if \code{TRUE} performs a series of analysis to assess the clonal landscape
-#'                              and returns a \link{ScoperClones} object. If \code{FALSE} then
+#'                              and returns a \link{ScoperClones} object. If \code{FALSE} (default) then
 #'                              a modified input \code{db} is returned. When grouping by \code{fields}, 
 #'                              \code{summarize_clones} should be \code{FALSE}.
-#'
 #' @return
-#' If \code{summarize_clones=TRUE} (default) a \link{ScoperClones} object is returned that includes the 
+#' If \code{summarize_clones=FALSE} (default) a modified \code{data.frame} is returned with clone identifiers in the 
+#' specified \code{clone} column.
+#' If \code{summarize_clones=TRUE} a \link{ScoperClones} object is returned that includes the 
 #' clonal assignment summary information and a modified input \code{db} in the \code{db} slot that 
 #' contains clonal identifiers in the specified \code{clone} column.
-#' If \code{summarize_clones=FALSE} modified \code{data.frame} is returned with clone identifiers in the 
-#' specified \code{clone} column.
 #'
 #' @details
 #' If \code{method="novj"}, then clonal relationships are inferred using an adaptive 
@@ -1099,7 +1136,9 @@ hierarchicalClones <- function(db, threshold, method=c("nt", "aa"), linkage=c("s
 #' db <- subset(ExampleDb, c_call == "IGHG")
 #' 
 #' # Find clonal groups
-#' results <- spectralClones(db, method="novj", germline="germline_alignment_d_mask")
+#' results <- spectralClones(db, method="novj", 
+#'            germline="germline_alignment_d_mask", 
+#'            summarize_clones=TRUE)
 #' 
 #' # Retrieve modified input data with clonal clustering identifiers
 #' df <- as.data.frame(results)
@@ -1110,12 +1149,11 @@ hierarchicalClones <- function(db, threshold, method=c("nt", "aa"), linkage=c("s
 #' @export
 spectralClones <- function(db, method=c("novj", "vj"), germline="germline_alignment", sequence="sequence_alignment",
                            junction="junction", v_call="v_call", j_call="j_call", clone="clone_id", fields=NULL,
-                           cell_id=NULL, locus="locus", only_heavy=TRUE, split_light=TRUE,
+                           cell_id=NULL, locus="locus", only_heavy=TRUE, split_light=FALSE,
                            targeting_model=NULL, len_limit=NULL, first=FALSE, cdr3=FALSE, mod3=FALSE, max_n=0, 
                            threshold=NULL, base_sim=0.95, iter_max=1000,  nstart=1000, nproc=1,
-                           verbose=FALSE, log=NULL,
-                           summarize_clones=TRUE) {
-    
+                           verbose=FALSE, log=NULL, summarize_clones=FALSE) {
+
     results <- defineClonesScoper(db = db, method = match.arg(method), model = "spectral", 
                                   germline = germline, sequence = sequence,
                                   junction = junction, v_call = v_call, j_call = j_call, clone = clone, fields = fields,
@@ -1124,8 +1162,7 @@ spectralClones <- function(db, method=c("novj", "vj"), germline="germline_alignm
                                   first = first, cdr3 = cdr3, mod3 = mod3, max_n = max_n,
                                   threshold = threshold, base_sim = base_sim,
                                   iter_max = iter_max, nstart = nstart, nproc = nproc,
-                                  verbose = verbose, log = log,
-                                  summarize_clones = summarize_clones)
+                                  verbose = verbose, log = log, summarize_clones = summarize_clones)
     
     # return results
     if (summarize_clones) {
@@ -1149,13 +1186,14 @@ defineClonesScoper <- function(db,
                                linkage = c("single", "average", "complete"), normalize = c("len", "none"),
                                germline = "germline_alignment", sequence = "sequence_alignment",
                                junction = "junction", v_call = "v_call", j_call = "j_call", clone = "clone_id",
-                               fields = NULL, cell_id = NULL, locus = NULL, only_heavy = TRUE, split_light = TRUE,
+                               fields = NULL, cell_id = NULL, locus = NULL, only_heavy = TRUE, split_light = FALSE,
                                targeting_model = NULL, len_limit = NULL,
                                first = FALSE, cdr3 = FALSE, mod3 = FALSE, max_n = 0, 
                                threshold = NULL, base_sim = 0.95,
                                iter_max = 1000, nstart = 1000, nproc = 1,
                                verbose = FALSE, log = NULL,
-                               summarize_clones = TRUE) {
+                               summarize_clones = FALSE) {
+  
     ### get model
     model <- match.arg(model)
     
@@ -1167,6 +1205,21 @@ defineClonesScoper <- function(db,
         stop("'db' must be a data frame")
     }
     
+    if (is(db, "data.table")) {
+        db <- as.data.frame(db)
+    }
+
+    if (!only_heavy) {
+      warning('only_heavy = FALSE is deprecated. Running as if only_heavy = TRUE')
+      only_heavy <- TRUE
+    }
+    
+    if (split_light) {
+      warning(paste('split_light = TRUE is deprecated. Please use split_light = FALSE.',
+                    'After clonal identification, light chain groups can be found with dowser::resolveLightChains'))
+      split_light <- FALSE
+    }
+
     ### check model and method
     if (model == "identical") {
         if (!(method %in% c("nt", "aa"))) {
@@ -1222,7 +1275,7 @@ defineClonesScoper <- function(db,
     
     ### check general required columns
     columns <- c(junction, v_call, j_call, fields) 
-    check <- checkColumns(db, columns)
+    check <- alakazam::checkColumns(db, columns)
     if (is.character(check)) { 
         stop(check)
     }
@@ -1241,7 +1294,7 @@ defineClonesScoper <- function(db,
     if (!is.null(cell_id) & !is.null(locus)) {
         # Check required columns for single-cell mode
         columns <- c(cell_id, locus) #, fields
-        check <- checkColumns(db, columns)
+        check <- alakazam::checkColumns(db, columns)
         if (check != TRUE) { stop(check) }
         
         # make a temp cell_id column to keep cell_ids specific to each fields 
@@ -1249,7 +1302,7 @@ defineClonesScoper <- function(db,
             db$cell_id_temp <- db %>%
                 dplyr::group_by(!!!rlang::syms(fields)) %>% 
                 dplyr::group_indices() 
-            db$cell_id_temp <- stri_join(db[[cell_id]], db$cell_id_temp, sep="_")     
+            db$cell_id_temp <- stringi::stri_join(db[[cell_id]], db$cell_id_temp, sep="_")     
             cell_id <- "cell_id_temp"
         }
         
@@ -1291,15 +1344,16 @@ defineClonesScoper <- function(db,
     }
     
     ### prepare db
-    ### Creates the initial vjl groups to identify clonally related sequences
+    ### Creates the initial VJ groups to identify clonally related sequences
     ### using the heavy chain.
     ### The vj group is created with groupGenes, using heavy chaing v and j calls only
-    ### (only_heavy=T) or also considering the vj light chain call (only_heavy=F). 
-    ### Then an additional l group is added, based on the junction length, not
-    ### using thegroupGenes. As only the heavy chain data is used for cloning, 
-    ### only the heavy chain sequences' junction length matters. At this point,
-    ### single cell data has one heavy chain sequence per cell, and one cell can
-    ### only belong to a v+j+heavy-chain-junction-length group.
+    ### (only_heavy=T). Then an additional l group is added, based on the junction length, not
+    ### using the groupGenes function. As only the heavy chain data is used for cloning, 
+    ### only the heavy chain sequences' junction length matters. 
+    ### At this point, single cell data has one heavy chain sequence per cell, and 
+    ### one cell can only belong to a v+j+heavy-chain-junction-length group.
+    
+    # TODO: modify function prepare_db to not accept only_heavy parameter any more
     results_prep <- prepare_db(db = db, 
                                junction = junction, v_call = v_call, j_call = j_call,
                                first = first, cdr3 = cdr3, fields = fields,
@@ -1432,7 +1486,7 @@ defineClonesScoper <- function(db,
                              # return result from each proc
                              return(db_gp)
                              # *********************************************************************************
-                         }
+    }
     
     ### stop the cluster
     if (nproc > 1) { parallel::stopCluster(cluster) }
@@ -1487,9 +1541,9 @@ defineClonesScoper <- function(db,
                              clone_count = length(unique(!!rlang::sym(clone))),
                              clone_id = paste(unique(!!rlang::sym(clone)), collapse = ","))
         vjl_gps$v_call <- sapply(1:nrow(vjl_gps),
-                                 function(i){ stri_join(unique(stri_split_fixed(vjl_gps$v_call[i], ",")[[1]]), collapse=",") })
+                                 function(i){ stringi::stri_join(unique(stringi::stri_split_fixed(vjl_gps$v_call[i], ",")[[1]]), collapse=",") })
         vjl_gps$j_call <- sapply(1:nrow(vjl_gps),
-                                 function(i){ stri_join(unique(stri_split_fixed(vjl_gps$j_call[i], ",")[[1]]), collapse=",") })
+                                 function(i){ stringi::stri_join(unique(stringi::stri_split_fixed(vjl_gps$j_call[i], ",")[[1]]), collapse=",") })
         
         ### calculate inter and intra distances
         df_inter_intra <- calculateInterVsIntra(db = db_cloned,
@@ -1552,49 +1606,112 @@ defineClonesScoper <- function(db,
                     db_l[[clone]][db_l[[cell_id]] == cellid] <- unique(db_cloned[[clone]][db_cloned[[cell_id]] == cellid])
                 } 
             }
-          # bind heavy and light chain data.frames
+            # bind heavy and light chain data.frames
             stopifnot(all(names(db_cloned) == names(db_l)))
             db_cloned <- bind_rows(db_cloned, db_l)
+
+
             # split clones by light chains
-            if (split_light) {
-                clones <- unique(db_cloned[[clone]])
-                clones <- clones[!is.na(clones)]
-                #TODO: parallelize this for loop
-                for (cloneid in clones) {
-                    db_c <- dplyr::filter(db_cloned, !!rlang::sym(clone) == cloneid)
-                    if (length(unique(db_c[[cell_id]])) == 1) next()
-                    db_c[['junction_l']] <- stringi::stri_length(db_c[[junction]])
-                    # Create temporary fake v_call and j_call, to avoid grouping 
-                    # again using heavy chain gene calls. This matters if first=FALSE
-                    # and "linker" ambiguous calls were left out of the same cluster id
-                    # because of the distance threshold. The goal now is to divide the 
-                    # heavy chain clones using light chain info only.
-                    # TODO: this is probably inefficient. Test is groupGenes could handle
-                    # v_call=NULL and j_call=NULL. Or maybe add an option only_light.
-                    db_c[[v_call]][db_c[['locus']] %in% c("IGH", "TRB", "TRD")] <- "IGHV0"
-                    db_c[[j_call]][db_c[['locus']] %in% c("IGH", "TRB", "TRD")] <- "IGHJ0"
-                    db_c <- alakazam::groupGenes(data = db_c,
-                                       v_call = v_call,
-                                       j_call = j_call,
-                                       junc_len = 'junction_l', 
-                                       cell_id = cell_id,
-                                       locus = locus,
-                                       only_heavy = FALSE,
-                                       first = FALSE)
-                    if (length(unique(db_c$vj_group)) == 1) next()
-                    db_c[[clone]] <- paste(db_c[[clone]], db_c$vj_group, sep="_") # TODO: IDs get connected with underscore, figure out when the underscore is removed
-                    for (cellid in unique(db_c[[cell_id]])) {
-                        db_cloned[[clone]][db_cloned[[clone]] == cloneid & db_cloned[[cell_id]] == cellid] <- 
-                            db_c[[clone]][db_c[[cell_id]] == cellid]
-                    }
-                }
-            }  
-            
+            # CGJ 1/28/25 -- based on group meeting this is no longer needed. 
+            # if (split_light) {
+            #     clones <- unique(db_cloned[[clone]])
+            #     clones <- clones[!is.na(clones)]
+            #     #TODO: test if substituting this for loop for the resolveLightChains function works
+            #     #TODO: option2: update groupGenes, add new param that allows for light chain gene grouping instead of heavy chain.
+            #     for (cloneid in clones) {
+            #         db_c <- dplyr::filter(db_cloned, !!rlang::sym(clone) == cloneid)
+            #         if (length(unique(db_c[[cell_id]])) == 1) next()
+            #         db_c[['junction_l']] <- stringi::stri_length(db_c[[junction]])
+            #         # Create temporary fake v_call and j_call, to avoid grouping
+            #         # again using heavy chain gene calls. This matters if first=FALSE
+            #         # and "linker" ambiguous calls were left out of the same cluster id
+            #         # because of the distance threshold. The goal now is to divide the
+            #         # heavy chain clones using light chain info only.
+            #         # TODO: this is probably inefficient. Test is groupGenes could handle
+            #         # v_call=NULL and j_call=NULL. Or maybe add an option only_light.
+            #         db_c[[v_call]][db_c[['locus']] %in% c("IGH", "TRB", "TRD")] <- "IGHV0"
+            #         db_c[[j_call]][db_c[['locus']] %in% c("IGH", "TRB", "TRD")] <- "IGHJ0"
+            #         db_c <- groupGenes(data = db_c,
+            #                            v_call = v_call,
+            #                            j_call = j_call,
+            #                            junc_len = 'junction_l',
+            #                            cell_id = cell_id,
+            #                            locus = locus,
+            #                            only_heavy = TRUE,
+            #                            first = FALSE)
+            #         if (length(unique(db_c$vj_group)) == 1) next()
+            #         # CGJ 11/11/24
+            #         # use the dowser logic to check for nonpaired heavy chains and
+            #         # find the closest one via hamming distance (from resolveLightChains)
+            #         hc_df <- dplyr::filter(db_c, !!rlang::sym(locus) %in% c("IGH", "TRB", "TRD"))
+            #         lc_df <- dplyr::filter(db_c, !!rlang::sym(locus) %in% c("IGK", "IGL", "TRA", "TRG"))
+            #         # see if any HCs are no paired
+            #         hc_unpaired <- dplyr::setdiff(unique(hc_df[['cell_id']]), unique(lc_df[['cell_id']]))
+            #         if(length(hc_unpaired) > 0){
+            #           unpaired <- hc_df[hc_df[['cell_id']] %in% hc_unpaired,]
+            #           paired <- hc_df[!hc_df[['cell_id']] %in% hc_unpaired,]
+            #           for(indx in 1:nrow(unpaired)){
+            #             rating <- unlist(lapply(1:nrow(paired), function(x){
+            #               temp <- rbind(paired[x,], unpaired[indx,])
+            #               if(nchar(temp[['sequence']][1]) != nchar(temp[['sequence']][2])){
+            #                 temp <- alakazam::padSeqEnds(temp[['sequence']])
+            #                 value <- alakazam::seqDist(temp[1], temp[2])
+            #               } else{
+            #                 value <- alakazam::seqDist(temp[['sequence']][1], temp[['sequence']][2])
+            #               }
+            #               return(value)
+            #             }))
+            #             rating <- as.numeric(rating)
+            #             # row number of heavy chain only df with lowest seq dist
+            #             proper_index <- which(rating == min(rating))
+            #             if(length(proper_index) > 1){
+            #               # find the subgroups that belong to the lowest seq dists
+            #               subgroups <- paired$vj_group[proper_index]
+            #               if(length(unique(subgroups)) > 1){
+            #                 # if there is more than one subgroup find the subgroup sizes of the 
+            #                 # subgroups being considered
+            #                 subgroup_size <- data.frame(vj_group = unique(subgroups))
+            #                 subgroup_size$sizes <- unlist(lapply(1:nrow(subgroup_size), function(x){
+            #                   nrow(paired[paired$vj_group == subgroup_size$vj_group[x],])
+            #                 }))
+            #                 # if there is one subgroup that is the largest use it
+            #                 if(length(which(subgroup_size$sizes == max(subgroup_size$sizes))) == 1){
+            #                   proper_index_value <- subgroup_size$vj_group[
+            #                     which(subgroup_size$sizes == max(subgroup_size$sizes))]
+            #                 } else { 
+            #                   # if there are more than one subgroup with the same size use the lower number
+            #                   potential_subgroups <- subgroup_size$vj_group[
+            #                     which(subgroup_size$sizes == max(subgroup_size$sizes))]
+            #                   potential_subgroups <- sort(potential_subgroups)
+            #                   proper_index_value <- potential_subgroups[1]
+            #                 }
+            #               } else{
+            #                 proper_index_value <- paired$vj_group[proper_index[1]]
+            #               }
+            #             } else{
+            #               proper_index_value <- paired$vj_group[proper_index]
+            #             }
+            #             unpaired$vj_group[indx] <- proper_index_value
+            #           }
+            #           # update the df to have the new unpaired df in place of the old one
+            #           # then bind it all together
+            #           hc_df <- rbind(paired, unpaired)
+            #         }
+            #         db_c <- rbind(hc_df, lc_df)
+            #         # CGJ end of logic update for unpaired heavy chains
+            #         db_c[[clone]] <- paste(db_c[[clone]], db_c$vj_group, sep="_") # TODO: IDs get connected with underscore, figure out when the underscore is removed
+            #         for (cellid in unique(db_c[[cell_id]])) {
+            #             db_cloned[[clone]][db_cloned[[clone]] == cloneid & db_cloned[[cell_id]] == cellid] <-
+            #                 db_c[[clone]][db_c[[cell_id]] == cellid]
+            #         }
+            #     }
+            # }
+
             # sort clone ids
             na.count <- sum(is.na(db_cloned[[clone]]))
             if (na.count > 0) {
                 db_na <- db_cloned[is.na(db_cloned[[clone]]), ]
-                db_cloned <- db_cloned[!is.na(db_cloned[[clone]]), ]        
+                db_cloned <- db_cloned[!is.na(db_cloned[[clone]]), ]
             }
             db_cloned$clone_temp <- db_cloned %>%
                 dplyr::group_by(!!rlang::sym(clone)) %>%
@@ -1607,8 +1724,9 @@ defineClonesScoper <- function(db,
                 db_cloned <- bind_rows(db_cloned, db_na)
             }
         }
-        
     }
+    
+    
     if (!is.null(fields) & single_cell) { db_cloned[[cell_id]] <- NULL}
     
     # return results
@@ -1769,9 +1887,21 @@ hierarchicalClones_helper <- function(db_gp,
     if (normalize == "len") {
         # calculate normalization factor
         junc_length <- unique(stringi::stri_length(seqs_unq))
-        hc <- hclust(as.dist(dist_mtx/junc_length), method = linkage)    
+        if(n_unq<65536){
+            hc <- stats::hclust(as.dist(dist_mtx/junc_length), method = linkage) 
+        }
+        else{
+            print(paste0("VJL group size: ", n_unq, ". Function hclust from fastcluster will be used for large vjl group."))
+            hc <- fastcluster::hclust(as.dist(dist_mtx/junc_length), method = linkage)
+        }
     } else if (normalize == "none") {
-        hc <- stats::hclust(as.dist(dist_mtx), method = linkage)    
+        if(n_unq<65536){
+            hc <- stats::hclust(as.dist(dist_mtx), method = linkage) 
+        }
+        else{
+            print(paste0("VJL group size: ", n_unq, ". Function hclust from fastcluster will be used for large vjl group."))
+            hc <- fastcluster::hclust(as.dist(dist_mtx), method = linkage)
+        }    
     }
     
     # cut the tree